(PC-17806) refactor(auth): extract all functions from api file
