+++ src/features/auth/api/__mocks__/useDepositAmountsByAge.ts
@@ -1,11 +1,3 @@
-export const signIn = jest.fn()
-
-export const useSignIn = jest.fn(() => signIn)
-
-export const signUp = jest.fn()
-
-export const useSignUp = jest.fn(() => signUp)
-
 export const useDepositAmountsByAge = jest.fn().mockReturnValue({
   fifteenYearsOldDeposit: '20\u00a0€',
   sixteenYearsOldDeposit: '30\u00a0€',
