+++ src/features/auth/constants.ts
@@ -1,0 +1,1 @@
+export const SIGNUP_NUMBER_OF_STEPS = 4 // email, password, birthday, cgu
