+++ src/features/auth/signup/SignupForm.tsx
@@ -3,8 +3,9 @@ import { StackScreenProps } from '@react-navigation/stack'
 import React, { FunctionComponent, useState, useCallback } from 'react'
 import { Keyboard } from 'react-native'
 
-import { SIGNUP_NUMBER_OF_STEPS, useSignUp } from 'features/auth/api'
+import { useSignUp } from 'features/auth/api/useSignUp'
 import { QuitSignupModal } from 'features/auth/components/QuitSignupModal'
+import { SIGNUP_NUMBER_OF_STEPS } from 'features/auth/constants'
 import { RootStackParamList } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
