+++ src/features/auth/signup/VerifyEligiblity/tests/VerifyEligibility.native.test.tsx
@@ -9,8 +9,6 @@ import { render, fireEvent } from 'tests/utils'
 
 import { VerifyEligibility } from '../VerifyEligibility'
 
-jest.mock('features/auth/api')
-
 jest.mock('features/navigation/helpers')
 jest.mock('features/navigation/navigationRef')
 jest.mock('react-query')
