+++ src/features/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -4,7 +4,7 @@ import QRCode from 'react-native-qrcode-svg'
 import styled from 'styled-components/native'
 
 import { CallToActionIcon, CategoryIdEnum, PopOverIcon, VenueTypeCodeKey } from 'api/gen/api'
-import { SIGNUP_NUMBER_OF_STEPS } from 'features/auth/api'
+import { SIGNUP_NUMBER_OF_STEPS } from 'features/auth/constants'
 import { ThreeShapesTicket } from 'features/bookings/components/ThreeShapesTicket'
 import { Icons } from 'features/cheatcodes/pages/AppComponents/Icons'
 import { Illustrations } from 'features/cheatcodes/pages/AppComponents/Illustrations'
