+++ src/features/profile/pages/DeleteProfile/ConfirmDeleteProfile.tsx
@@ -2,7 +2,7 @@ import { useNavigation } from '@react-navigation/native'
 import React from 'react'
 import styled from 'styled-components/native'
 
-import { useAccountSuspend } from 'features/auth/api'
+import { useAccountSuspend } from 'features/auth/api/useAccountSuspend'
 import { useLogoutRoutine } from 'features/auth/logout/useLogoutRoutine'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
