+++ src/features/user/helpers/useGetDepositAmountsByAge/useGetDepositAmountsByAge.test.ts
@@ -7,8 +7,6 @@ import { useGetDepositAmountsByAge } from './useGetDepositAmountsByAge'
 const TODAY = '2022-10-24'
 mockdate.set(new Date(TODAY))
 
-jest.mock('features/auth/api')
-
 describe('useGetDepositAmountsByAge', () => {
   it('should return nothing when birthDate is not defined', () => {
     const { result } = renderHook(() => useGetDepositAmountsByAge())
